split apps.plugin into multiple files
