+++ src/collectors/apps.plugin/apps_proc_pids.c
@@ -1,0 +1,147 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "apps_plugin.h"
+
+static inline struct pid_stat *get_pid_entry(pid_t pid) {
+    if(unlikely(all_pids[pid]))
+        return all_pids[pid];
+
+    struct pid_stat *p = callocz(sizeof(struct pid_stat), 1);
+    p->fds = mallocz(sizeof(struct pid_fd) * MAX_SPARE_FDS);
+    p->fds_size = MAX_SPARE_FDS;
+    init_pid_fds(p, 0, p->fds_size);
+    p->pid = pid;
+
+    DOUBLE_LINKED_LIST_APPEND_ITEM_UNSAFE(root_of_pids, p, prev, next);
+
+    all_pids[pid] = p;
+    all_pids_count++;
+
+    return p;
+}
+
+static inline void del_pid_entry(pid_t pid) {
+    struct pid_stat *p = all_pids[pid];
+
+    if(unlikely(!p)) {
+        netdata_log_error("attempted to free pid %d that is not allocated.", pid);
+        return;
+    }
+
+    debug_log("process %d %s exited, deleting it.", pid, p->comm);
+
+    DOUBLE_LINKED_LIST_REMOVE_ITEM_UNSAFE(root_of_pids, p, prev, next);
+
+    // free the filename
+#ifndef __FreeBSD__
+    {
+        size_t i;
+        for(i = 0; i < p->fds_size; i++)
+            if(p->fds[i].filename)
+                freez(p->fds[i].filename);
+    }
+#endif
+    freez(p->fds);
+
+    freez(p->fds_dirname);
+    freez(p->stat_filename);
+    freez(p->status_filename);
+    freez(p->limits_filename);
+#ifndef __FreeBSD__
+    arl_free(p->status_arl);
+#endif
+    freez(p->io_filename);
+    freez(p->cmdline_filename);
+    freez(p->cmdline);
+    freez(p);
+
+    all_pids[pid] = NULL;
+    all_pids_count--;
+}
+
+int collect_data_for_pid(pid_t pid, void *ptr) {
+    if(unlikely(pid < 0 || pid > pid_max)) {
+        netdata_log_error("Invalid pid %d read (expected %d to %d). Ignoring process.", pid, 0, pid_max);
+        return 0;
+    }
+
+    struct pid_stat *p = get_pid_entry(pid);
+    if(unlikely(!p || p->read)) return 0;
+    p->read = true;
+
+    // debug_log("Reading process %d (%s), sortlist %d", p->pid, p->comm, p->sortlist);
+
+    // --------------------------------------------------------------------
+    // /proc/<pid>/stat
+
+    if(unlikely(!managed_log(p, PID_LOG_STAT, read_proc_pid_stat(p, ptr))))
+        // there is no reason to proceed if we cannot get its status
+        return 0;
+
+    // check its parent pid
+    if(unlikely(p->ppid < 0 || p->ppid > pid_max)) {
+        netdata_log_error("Pid %d (command '%s') states invalid parent pid %d. Using 0.", pid, p->comm, p->ppid);
+        p->ppid = 0;
+    }
+
+    // --------------------------------------------------------------------
+    // /proc/<pid>/io
+
+    managed_log(p, PID_LOG_IO, read_proc_pid_io(p, ptr));
+
+    // --------------------------------------------------------------------
+    // /proc/<pid>/status
+
+    if(unlikely(!managed_log(p, PID_LOG_STATUS, read_proc_pid_status(p, ptr))))
+        // there is no reason to proceed if we cannot get its status
+        return 0;
+
+    // --------------------------------------------------------------------
+    // /proc/<pid>/fd
+
+    if(enable_file_charts) {
+        managed_log(p, PID_LOG_FDS, read_pid_file_descriptors(p, ptr));
+        managed_log(p, PID_LOG_LIMITS, read_proc_pid_limits(p, ptr));
+    }
+
+    // --------------------------------------------------------------------
+    // done!
+
+    if(unlikely(debug_enabled && include_exited_childs && all_pids_count && p->ppid && all_pids[p->ppid] && !all_pids[p->ppid]->read))
+        debug_log("Read process %d (%s) sortlisted %d, but its parent %d (%s) sortlisted %d, is not read", p->pid, p->comm, p->sortlist, all_pids[p->ppid]->pid, all_pids[p->ppid]->comm, all_pids[p->ppid]->sortlist);
+
+    // mark it as updated
+    p->updated = true;
+    p->keep = false;
+    p->keeploops = 0;
+
+    return 1;
+}
+
+void cleanup_exited_pids(void) {
+    size_t c;
+    struct pid_stat *p = NULL;
+
+    for(p = root_of_pids; p ;) {
+        if(!p->updated && (!p->keep || p->keeploops > 0)) {
+            if(unlikely(debug_enabled && (p->keep || p->keeploops)))
+                debug_log(" > CLEANUP cannot keep exited process %d (%s) anymore - removing it.", p->pid, p->comm);
+
+            for(c = 0; c < p->fds_size; c++)
+                if(p->fds[c].fd > 0) {
+                    file_descriptor_not_used(p->fds[c].fd);
+                    clear_pid_fd(&p->fds[c]);
+                }
+
+            pid_t r = p->pid;
+            p = p->next;
+            del_pid_entry(r);
+        }
+        else {
+            if(unlikely(p->keep)) p->keeploops++;
+            p->keep = false;
+            p = p->next;
+        }
+    }
+}
+
